+++ src/Autocomplete/translations/AutocompleteBundle.sr_RS.php
@@ -13,4 +13,5 @@ return [
     'Loading more results...' => 'Учитавање још резултата...',
     'No results found' => 'Nema rezultata',
     'No more results' => 'Nema više rezultata',
+    // 'Add %placeholder%...' => 'Add %placeholder%...',
 ];