+++ src/Autocomplete/translations/AutocompleteBundle.uk.php
@@ -13,4 +13,5 @@ return [
     'Loading more results...' => 'Завантаження додаткових результатів...',
     'No results found' => 'Нічого не знайдено',
     'No more results' => 'Більше результатів немає',
+    // 'Add %placeholder%...' => 'Add %placeholder%...',
 ];