+++ src/Svelte/assets/vitest.config.mjs
@@ -1,7 +1,7 @@
 // vitest.config.ts
 import { defineConfig, mergeConfig } from 'vitest/config';
 import { svelte } from '@sveltejs/vite-plugin-svelte';
-import configShared from '../../../vitest.config.js'
+import configShared from '../../../vitest.config.mjs'
 
 export default mergeConfig(
     configShared,