+++ src/Turbo/assets/vitest.config.mjs
@@ -1,5 +1,5 @@
 import { defineConfig, mergeConfig } from 'vitest/config';
-import configShared from '../../../vitest.config.js'
+import configShared from '../../../vitest.config.mjs'
 import path from 'path';
 
 export default mergeConfig(