+++ ux.symfony.com/src/Twig/Alert.php
@@ -11,7 +11,6 @@
 
 namespace App\Twig;
 
-use App\Service\UxPackageRepository;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
 #[AsTwigComponent]
@@ -20,20 +19,11 @@ class Alert
     public string $type = 'success';
     public string $message;
 
-    public function __construct(private UxPackageRepository $packageRepository)
-    {
-    }
-
     public function getIcon(): string
     {
         return match ($this->type) {
-            'success' => 'circle-check',
-            'danger' => 'circle-exclamation',
+            'success' => 'bi:check-circle',
+            'danger' => 'bi:exclamation-circle',
         };
-    }
-
-    public function getPackageCount(): int
-    {
-        return $this->packageRepository->count();
     }
 }