+++ src/Icons/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.20.0
+
+-   Add `aliases` configuration option to define icon alternative names.
+-   Add support for `int` and `float` attribute values in `<twig:ux:icon />`.
+
 ## 2.19.0
 
 -   Add `ignore_not_found` option to silence error during rendering if the